requirements.txt
mistralai==1.2.3
openai==1.56.1
pydantic==2.10.3
PyYAML==6.0.2
setuptools==75.6.0